refactor: 컴포넌트 계층 축소 Tabs->Form->List
